refact : 인터뷰 결과 상세보기 분리
