+++ client/pages/api/interview-feedback/index.ts
@@ -32,17 +32,30 @@ export default async function handler(
     }
 
     const prompt = `
-    당신은 기술 면접 피드백을 제공하는 전문가입니다. 다음 답변의 강점과 개선점을 분석하여 간결하게 JSON 형태로만 응답하세요.
+      목적/작업 유형: 면접 답변 피드백
 
-    다음은 기술 면접 질문과 답변입니다. 답변의 강점과 개선점을 분석하여 JSON 형태로 제공해주세요.
-    강점과 개선점은 각각 2~6개로 구성해주세요. 코드 예시에 대한 내용은 포함되지 않습니다.
+      역할 및 과제: 경험 많은 면접관으로서 음성 답변${question}에 대한 평가${answer}를 실시
 
-    질문: ${question}
-    답변: ${answer}
+      평가 기준:
+      1. 기술적 정확성
+      2. 개념 이해도
+      3. 설명의 명확성
+      4. 실무 적용성
+      5. 구조화된 답변
 
-    다음 JSON 형식으로만 응답해주세요:
-      "strengths": ["강점1", "강점2", "..."],
-      "improvements": ["개선점1", "개선점2", "..."]
+      응답 형식:
+      {
+        "overall_rating":  "탁월함|우수함|양호함|기본적|심각함"
+        "strengths": ["강점1", "강점2", "..."],
+        "improvements": ["개선점1", "개선점2", "..."],
+      }
+
+      지침:
+      - {overall_rating, strengths, improvements} 필드만 포함한 JSON 형식으로만 응답합니다.
+      - 음성 응답 상황을 고려하여 코드 작성 대신 논리적 설명 능력에 중점을 둡니다.
+      - 강점은 2-5개, 개선점은 0-5개 제시합니다.
+      - 탁월함 답변엔 improvements 첫항목: "훌륭한 답변입니다. 추가 고려사항:"
+      - 면접에서 실제로 중요한 요소(예: 문제 해결 접근법, 기술 선택의 이유 등)에 초점을 맞춥니다.
     `;
 
     const response = await anthropic.messages.create({
