+++ client/src/features/interview/screen-result/ui/content-detail/RecordingDetails.tsx
@@ -1,6 +1,7 @@
 import { styled } from 'styled-components';
+
 import { styleMixin, V } from '@/src/shared/styles';
-import { keys } from '@/src/shared/utils';
+import FeedbackResult from './FeedbackResult';
 
 type RecordContent = {
   script: string;
@@ -15,6 +16,29 @@ type Props = {
   recordContents: RecordContent;
 };
 
+type ContentProps = {
+  title: string;
+  children: React.ReactNode;
+};
+
+function Section({ title, children }: ContentProps) {
+  return (
+    <ContentSection>
+      <SectionHeader>
+        <Label>{title}</Label>
+        <Divider />
+      </SectionHeader>
+      {children}
+    </ContentSection>
+  );
+}
+
+/**
+ * 유저의 녹화 정보 상세 결과
+ *
+ * @param question - 녹화 질문
+ * @param recordContent - 유저가 입력한 자료들
+ */
 export default function RecordingDetails({ question, recordContents }: Props) {
   const {
     script,
@@ -31,68 +55,27 @@ export default function RecordingDetails({ question, recordContents }: Props) {
         <Timer>{timer || '00:00'}</Timer>
       </TimerSection>
 
-      <ContentSection>
-        <SectionHeader>
-          <Label>인터뷰 질문</Label>
-          <Divider />
-        </SectionHeader>
+      <Section title="인터뷰 질문">
         <Content>{question}</Content>
-      </ContentSection>
+      </Section>
 
-      <ContentSection>
-        <SectionHeader>
-          <Label>스크립트</Label>
-          <Divider />
-        </SectionHeader>
+      <Section title="스크립트">
         <ContentCard>{script || '작성한 대본이 없습니다.'}</ContentCard>
-      </ContentSection>
+      </Section>
 
-      <ContentSection>
-        <SectionHeader>
-          <Label>음성 스크립트</Label>
-          <Divider />
-        </SectionHeader>
-        <ContentCard>{voiceScript || '발음 인식이 진행되지 않았습니다.'}</ContentCard>
-      </ContentSection>
+      <Section title="음성 스크립트">
+        <ContentCard>
+          {voiceScript || '발음 인식이 진행되지 않았습니다.'}
+        </ContentCard>
+      </Section>
 
       <FeedbackSection>
         <SectionHeader>
           <Label>피드백 요청 결과</Label>
           <Divider />
         </SectionHeader>
-
-        <FeedbackContainer>
-          {strengths.length > 0 && (
-            <FeedbackWrapper>
-              <CategoryLabel>
-                강점
-              </CategoryLabel>
-              {strengths.map((strength, idx) => (
-                <FeedbackItem $iconType="strengths" key={keys(strength, idx)}>
-                  {strength}
-                </FeedbackItem>
-              ))}
-            </FeedbackWrapper>
-          )}
-
-          {improvements.length > 0 && (
-            <FeedbackWrapper>
-              <CategoryLabel>
-                개선점
-              </CategoryLabel>
-              {improvements.map((improvement, idx) => (
-                <FeedbackItem
-                  $iconType="improvements"
-                  key={keys(improvement, idx)}
-                >
-                  {improvement}
-                </FeedbackItem>
-              ))}
-            </FeedbackWrapper>
-          )}
-        </FeedbackContainer>
+        <FeedbackResult strengths={strengths} improvements={improvements} />
       </FeedbackSection>
-
     </Wrapper>
   );
 }
@@ -183,7 +166,7 @@ const Timer = styled.p`
   font-size: 1.8rem;
   font-weight: 500;
   color: ${(props) => props.theme.colors.text};
-  font-family: 'Roboto Mono', monospace;
+  font-family: "Roboto Mono", monospace;
 
   @media screen and (max-width: ${V.mediaMobile}) {
     font-size: 1.4rem;
@@ -194,68 +177,4 @@ const FeedbackSection = styled(ContentSection)`
   width: 100%;
   margin-top: 0.8rem;
   padding-top: 1.2rem;
-`;
-
-const FeedbackContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  width: 100%;
-`;
-
-const FeedbackWrapper = styled.div`
-  margin-top: 0.5rem;
-  border-radius: 0.8rem;
-  padding: 1.2rem 1.5rem;
-`;
-
-const CategoryLabel = styled.div`
-  display: flex;
-  align-items: center;
-  font-size: 1.6rem;
-  font-weight: 500;
-  color: ${(props) => props.theme.colors.text};
-  margin-bottom: 1rem;
-  padding-bottom: 0.8rem;
-  border-bottom: 1px dashed ${(props) => props.theme.colors.mainLine};
-
-  @media screen and (max-width: ${V.mediaMobile}) {
-    font-size: 1.4rem;
-  }
-`;
-
-type FeedItemStyle = {
-  $iconType: 'improvements' | 'strengths';
-};
-
-const FeedbackItem = styled.p<FeedItemStyle>`
-  position: relative;
-  font-size: 1.7rem;
-  line-height: 1.5;
-  color: ${(props) => props.theme.colors.text};
-  margin: 0.8rem 0;
-  padding-left: 1.8rem;
-
-  &::before {
-    content: "";
-    position: absolute;
-    left: 0;
-    top: 0.6rem;
-    width: 0.8rem;
-    height: 0.8rem;
-    border-radius: 50%;
-    background-color: ${(props) =>
-    (props.$iconType === 'strengths' ? 'green' : 'orange')};
-  }
-
-  @media screen and (max-width: ${V.mediaMobile}) {
-    font-size: 1.4rem;
-    padding-left: 1.6rem;
-
-    &::before {
-      top: 0.5rem;
-      width: 0.7rem;
-      height: 0.7rem;
-    }
-  }
 `;
