+++ client/src/features/interview/screen-result/ui/content-detail/index.ts
@@ -1,0 +1,1 @@
+export { default as RecordingDetails } from './RecordingDetails';
